+++ packages/graphql/config/index.ts
@@ -1,0 +1,6 @@
+export * from './environment'
+
+export { default as loadSerializableQuery } from './loadSerializableQuery'
+export type * from './loadSerializableQuery'
+
+export { default as useSerializablePreloadedQuery } from './useSerializablePreloadedQuery'