+++ packages/graphql/index.ts
@@ -1,0 +1,6 @@
+export * from './config'
+
+export { default as RelayProvider } from './RelayProvider'
+
+export { default as withRelay } from './withRelay'
+export type * from './withRelay/types'