+++ packages/utils/functions/form/setFormApiErrors/index.ts
@@ -1,8 +1,16 @@
+import entries from 'lodash/entries'
+import head from 'lodash/head'
+import map from 'lodash/map'
 import { FieldValues, Path, UseFormReturn } from 'react-hook-form'
 
 export const setFormApiErrors = <T extends FieldValues>(form: UseFormReturn<T>, err: any) => {
-  Object.entries(err?.response ?? {}).map(
-    ([key, errors]) =>
-      form.setError(key as Path<T>, { type: 'manual', message: (errors as string[])[0] }), // we asume that there will be only one error per field
-  )
+  if (err.response?.data && typeof err.response.data === 'object') {
+    map(entries(err.response.data), ([key, errors]) => {
+      const fieldKey = key as Path<T>
+      const errorMessage = head(errors as string[])
+      if (!!form.getValues(fieldKey) && typeof errorMessage === 'string') {
+        form.setError(fieldKey, { type: 'manual', message: errorMessage })
+      }
+    })
+  }
 }